Allow text to occur on line after @tag. Concat with "\n".
These changes will allow the addition of an @example tag
with multiple lines of code. Concatting with "\n" rather than
a space is a better idea since this will be rendered in HTML
as a space anyway, and preserves formatting in some places
where it is important. If a non-HTML doclet needs space rather
than newlines, then it should be the responsiblity of the doclet
to simply perform a gsub("\n", " ")